refactor: restructure lint tool organizations
