+++ tools/lint/.covgate
@@ -1,1 +1,1 @@
-90.0
+30.0
