+++ tools/lint/linter/.covgate
@@ -1,0 +1,1 @@
+93.3
