+++ tools/lint/linter/analysis/.covgate
@@ -1,0 +1,1 @@
+97.8
