+++ tools/lint/linter/analysis/scanner.go
@@ -1,14 +1,6 @@
-package main
+package analysis
 
 import "strings"
-
-// ProseSpan represents a contiguous segment of prose text within a line.
-// StartCol is the 1-based byte offset of the span's first character in
-// the original line.
-type ProseSpan struct {
-	StartCol int
-	Text     string
-}
 
 type zone int
