+++ tools/lint/linter/analysis/scanner_test.go
@@ -1,4 +1,4 @@
-package main
+package analysis
 
 import (
 	"testing"
