+++ tools/lint/linter/componentstyle/.covgate
@@ -1,0 +1,1 @@
+100.0
