+++ tools/lint/linter/importblock/.covgate
@@ -1,0 +1,1 @@
+100.0
