+++ tools/lint/linter/importsorted/.covgate
@@ -1,0 +1,1 @@
+100.0
