+++ tools/lint/linter/importused/.covgate
@@ -1,0 +1,1 @@
+100.0
