+++ tools/lint/linter/nodoubledash/.covgate
@@ -1,0 +1,1 @@
+100.0
