+++ tools/lint/main_test.go
@@ -59,61 +59,3 @@ func TestFindContentRoot(t *testing.T) {
 		}
 	})
 }
-
-func TestLintFile(t *testing.T) {
-	t.Run("missing file returns error", func(t *testing.T) {
-		_, err := lintFile("/no/such/file.mdx", nil, nil)
-		if err == nil {
-			t.Error("expected error for missing file")
-		}
-	})
-
-	t.Run("no violations", func(t *testing.T) {
-		dir := t.TempDir()
-		path := filepath.Join(dir, "test.mdx")
-		content := []byte("# Title\n\nHello world.\n")
-		if err := os.WriteFile(path, content, 0o644); err != nil {
-			t.Fatal(err)
-		}
-		vs, err := lintFile(path, []Rule{NoDoubleDash{}}, nil)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(vs) != 0 {
-			t.Errorf("expected 0 violations, got %d: %v", len(vs), vs)
-		}
-	})
-
-	t.Run("prose rule violation", func(t *testing.T) {
-		dir := t.TempDir()
-		path := filepath.Join(dir, "test.mdx")
-		content := []byte("# Title\n\nBad -- prose.\n")
-		if err := os.WriteFile(path, content, 0o644); err != nil {
-			t.Fatal(err)
-		}
-		vs, err := lintFile(path, []Rule{NoDoubleDash{}}, nil)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(vs) != 1 {
-			t.Errorf("expected 1 violation, got %d: %v", len(vs), vs)
-		}
-	})
-
-	t.Run("file rule violation", func(t *testing.T) {
-		dir := t.TempDir()
-		path := filepath.Join(dir, "test.mdx")
-		content := "import Missing from '/snippets/missing.mdx';\n\nHello.\n"
-		if err := os.WriteFile(path, []byte(content), 0o644); err != nil {
-			t.Fatal(err)
-		}
-		fr := ImportResolvesRule{ContentRoot: dir}
-		vs, err := lintFile(path, nil, []FileRule{fr})
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if len(vs) != 1 {
-			t.Errorf("expected 1 violation, got %d: %v", len(vs), vs)
-		}
-	})
-}
